+++ RELEASE_NOTES.md
@@ -9,9 +9,9 @@ This version adds some iOS13/SwiftUI/Combine-specific logic.
 * `KeyboardGridRow` has been deleted and replaced with an `HStack`.
 * `KeyboardSetting` can resolve a unique key for an `Identifiable` context.
 * `PersistedKeyboardSetting` can resolve a unique key for an `Identifiable` context.
-* `View+KeyboardStyles` lets you apply the system keyboard styles to SwiftUI views.
 * `View+KeyboardActionGestures` lets you apply keyboard gestures for a certain action to SwiftUI views.
 * `View+KeyboardGestures` lets you apply keyboard gestures to SwiftUI views.
+* `View+SystemKeyboardButton` lets you apply the system keyboard styles to SwiftUI views.
 
 Breaking changes: